Move `WhisperModel` to `enums` and lazy-import heavy deps in CLI
